partitions main source file [clang]
COMMENTS:
Moves the methods of the vector class into their own source file,
as it should be.

Makes the grow() utility static since it needs not to be
called from elsewhere just from the push_back() method.

Adds a header for the methods of the vector class - Vector.h.
